refactor(webapp): normalize log event types
