+++ cmd/webapp/src/stores/logs.ts
@@ -2,7 +2,7 @@
  * Logs state store
  */
 import { writable } from 'svelte/store';
-import type { LogEvent } from '../types/stores';
+import type { LogEvent } from '../types/logs';
 
 export const logEvents = writable<LogEvent[]>([]);
 export const logsRetryCount = writable<number>(0);
