+++ cmd/webapp/src/types/run.ts
@@ -1,0 +1,9 @@
+/**
+ * Types related to the run view and environment configuration.
+ */
+
+export interface EnvRow {
+    id: number;
+    key: string;
+    value: string;
+}
